improvements

+++ src/controllers/comment.controller.js
@@ -1,4 +1,4 @@
-const { onSuccess, onCreated, onDeleted, onError } = require('../_shared/handlers/index');
+const { onSuccess, onCreated, onDeleted, onError, onUnathorized, onNotFound } = require('../_shared/handlers/index');
 
 const commentService = require('../services/comment.service');
 
@@ -21,6 +21,10 @@ class Controller {
             const { body } = ctx.request;
 
             const created = await commentService.update(id, body);
+
+            if (post.author !== ctx.postId) {
+                return onUnathorized(ctx, 'Only the comment original author can update!');
+            }
 
             onSuccess(ctx, created);
         } catch (err) {
@@ -46,6 +50,10 @@ class Controller {
 
             const comment = await commentService.getById(id);
 
+            if (!comment) {
+                return onNotFound(ctx, 'The comment you are looking for does not exists or was removed by author!');
+            }
+
             onSuccess(ctx, comment);
         } catch (err) {
             onError(ctx, err.message);
@@ -55,6 +63,10 @@ class Controller {
     async remove(ctx) {
         try {
             const { id } = ctx.params;
+
+            if (post.author !== ctx.postId) {
+                return onUnathorized(ctx, 'Only the comment original author can delete!');
+            }
 
             await commentService.remove(id);
